+++ src/cpp/galay-mcp/common/mcp_json.h
@@ -12,7 +12,7 @@
 #define GALAY_MCP_COMMON_MCPJSON_H
 
 #include "mcp_error.h"
-#include <simdjson.h>
+#include <galay/thirdparty/simdjson/simdjson.h>
 #include <expected>
 #include <memory>
 #include <string>